requirements.txt
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
aiounittest==1.4.2
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.0
black==22.6.0
charset-normalizer==2.1.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.79.0
fastapi-utils==0.2.1
frozenlist==1.3.1
GeoAlchemy2==0.12.3
greenlet==1.1.2
h11==0.13.0
idna==3.3
jose==1.0.0
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
pyasn1==0.4.8
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
rsa==4.9
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
utm==0.7.0
uvicorn==0.18.2
websocket-client==1.4.0
websockets==10.3
wrapt==1.14.1
yarl==1.8.1
